# packages/contracts/openapi/agent-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Course Tutor Agent API",
    "version": "1.0.0",
    "description": "Canonical v1 contract. x-implementation-status distinguishes implemented, partial and planned operations."
  },
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/v1/admin/programmes": {
      "post": {
        "operationId": "createProgramme",
        "x-implementation-status": "implemented",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProgrammeRequest"}}}},
        "responses": {"201": {"description": "Programme registered"}, "403": {"$ref": "#/components/responses/Forbidden"}}
      }
    },
    "/v1/admin/courses": {
      "post": {
        "operationId": "createCourse",
        "x-implementation-status": "implemented",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCourseRequest"}}}},
        "responses": {"201": {"description": "Course and course run registered"}, "403": {"$ref": "#/components/responses/Forbidden"}}
      }
    },
    "/v1/admin/courses/{course_id}/ingestions": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}],
      "post": {
        "operationId": "createCourseIngestion",
        "x-implementation-status": "implemented",
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateIngestionRequest"}}}},
        "responses": {"202": {"description": "Version build queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestionJob"}}}}, "403": {"$ref": "#/components/responses/Forbidden"}}
      }
    },
    "/v1/admin/ingestions/{job_id}": {
      "parameters": [{"$ref": "#/components/parameters/JobId"}],
      "get": {
        "operationId": "getIngestion",
        "x-implementation-status": "implemented",
        "responses": {"200": {"description": "Ingestion state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestionJob"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/v1/admin/ingestions/{job_id}/retry": {
      "parameters": [{"$ref": "#/components/parameters/JobId"}],
      "post": {
        "operationId": "retryIngestion",
        "x-implementation-status": "implemented",
        "responses": {
          "200": {"description": "Dead-lettered ingestion reset and queued"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"description": "Completed ingestion cannot be retried"}
        }
      }
    },
    "/v1/admin/courses/{course_id}/sources": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}],
      "get": {
        "operationId": "listCourseSources",
        "x-implementation-status": "implemented",
        "responses": {"200": {"description": "Source and extraction status list"}, "403": {"$ref": "#/components/responses/Forbidden"}}
      }
    },
...
    "/v1/admin/courses/{course_id}/versions/{version_id}/publish": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}, {"$ref": "#/components/parameters/VersionId"}],
      "post": {
        "operationId": "publishContentVersion",
        "x-implementation-status": "implemented",
        "responses": {"200": {"description": "Version published and active alias updated"}, "409": {"description": "Version is not ready"}}
      }
    },
    "/v1/admin/courses/{course_id}/versions/{version_id}/rollback": {
      "parameters": [{"$ref": "#/components/parameters/CourseId"}, {"$ref": "#/components/parameters/VersionId"}],
      "post": {
        "operationId": "rollbackContentVersion",
        "x-implementation-status": "implemented",
        "responses": {"200": {"description": "Previous valid version activated"}, "409": {"description": "No valid rollback target"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "CourseId": {"name": "course_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "JobId": {"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
      "VersionId": {"name": "version_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
    "responses": {
...
      "Forbidden": {"description": "Insufficient role, membership or tenant scope"},
      "NotFound": {"description": "Resource not found in the authorized scope"}
    },
    "schemas": {
...
      "CreateProgrammeRequest": {"type": "object", "additionalProperties": false, "required": ["code", "name"], "properties": {"code": {"type": "string", "minLength": 1, "maxLength": 64}, "name": {"type": "string", "minLength": 1, "maxLength": 255}}},
      "CreateCourseRequest": {"type": "object", "additionalProperties": false, "required": ["programme_id", "code", "name", "level", "run_key", "source_path"], "properties": {"programme_id": {"type": "string", "format": "uuid"}, "code": {"type": "string"}, "name": {"type": "string"}, "level": {"enum": ["primary", "middle_school", "high_school", "undergraduate", "postgraduate"]}, "run_key": {"type": "string"}, "source_path": {"type": "string"}, "scan_interval_seconds": {"type": "integer", "minimum": 30, "default": 900}}},
      "CreateIngestionRequest": {"type": "object", "additionalProperties": false, "properties": {"reason": {"type": "string", "maxLength": 500}, "force_rebuild": {"type": "boolean", "default": false}}},
      "IngestionJob": {"type": "object", "required": ["job_id", "course_id", "status"], "properties": {"job_id": {"type": "string", "format": "uuid"}, "course_id": {"type": "string", "format": "uuid"}, "version_id": {"type": ["string", "null"], "format": "uuid"}, "status": {"enum": ["pending", "building", "ready", "published", "failed", "dead_lettered"]}, "last_error": {"type": ["string", "null"]}}},
...
    }
  }
}